pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-evals"
version = "0.1.0"
requires-python = ">=3.9"
readme = "README.md"
dependencies = [
    "anthropic",
    "backoff",
    "blobfile",
    "datasets",
    "filelock",
    "fire",
    "google-genai",
    "jinja2",
    "matplotlib",
    "numpy",
    "ollama",
    "openai>=1.0.0",
    "pandas",
    "python-dotenv",
    "pyyaml",
    "requests",
    "scipy",
    "tabulate",
    "tqdm",
]

[tool.setuptools]
packages = ["simple_evals", "simple_evals.sampler"]

[tool.setuptools.package-dir]
simple_evals = "."

[tool.setuptools.package-data]
simple_evals = ["*.py"]
